move a test that imports gempy out of astrodata and into gempy where the function being tested lives.
